requirements.txt
torch==2.0.1
torchvision==0.15.2
ultralytics>=8.0.0
numpy<2.0
scipy>=1.7.0
opencv-python>=4.5.0
Pillow>=9.0.0
pandas>=1.3.0
matplotlib>=3.3.0
seaborn>=0.11.0
requests>=2.25.0
pathlib2>=2.3.0
tqdm>=4.60.0 